+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add supported glossary languages: Portuguese (`'PT'`), Russian (`'RU'`), and Chinese (`'ZH'`).
 * Add `output_format` parameter for document upload, to specify the desired file
   format of the output file
+* Add new language Arabic (`'AR'`).
 
 
 ## [1.11.0] - 2023-07-06