+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add `output_format` parameter for document upload, to specify the desired file
   format of the output file
 * Add new language Arabic (`'AR'`).
+### Security
+* Run `npm audit fix` to update dependencies.
 
 
 ## [1.11.0] - 2023-07-06