+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+* Add supported glossary languages: Danish (`'da'`), Norwegian (bokmål)
+  (`'nb'`), and Swedish (`'sv'`).
+
+
 ## [1.12.0] - 2024-04-08
 ### Added
 * Verify keepalive is used by clients by rejecting session reuse on new sockets,
@@ -176,6 +182,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial version.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-mock/compare/v1.12.0...HEAD
 [1.12.0]: https://github.com/DeepLcom/deepl-mock/compare/v1.11.0...v1.12.0
 [1.11.0]: https://github.com/DeepLcom/deepl-mock/compare/v1.10.0...v1.11.0
 [1.10.0]: https://github.com/DeepLcom/deepl-mock/compare/v1.9.1...v1.10.0