+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [1.13.0] - 2024-09-13
 ### Added
 * Add supported glossary languages: Danish (`'da'`), Norwegian (bokmål)
   (`'nb'`), and Swedish (`'sv'`).
@@ -184,7 +184,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial version.
 
 
-[Unreleased]: https://github.com/DeepLcom/deepl-mock/compare/v1.12.0...HEAD
+[1.13.0]: https://github.com/DeepLcom/deepl-mock/compare/v1.12.0...v1.13.0
 [1.12.0]: https://github.com/DeepLcom/deepl-mock/compare/v1.11.0...v1.12.0
 [1.11.0]: https://github.com/DeepLcom/deepl-mock/compare/v1.10.0...v1.11.0
 [1.10.0]: https://github.com/DeepLcom/deepl-mock/compare/v1.9.1...v1.10.0